+++ marimo/_server/api/rename.py
@@ -7,9 +7,10 @@ from dataclasses import dataclass
 import tornado.web
 
 from marimo import _loggers
-from marimo._server import sessions
+from marimo._server import server_utils as sessions
 from marimo._server.api import status
 from marimo._server.api.validated_handler import ValidatedHandler
+from marimo._server.sessions import get_manager
 from marimo._server.utils import canonicalize_filename
 from marimo._utils.parse_dataclass import parse_raw
 
@@ -26,7 +27,7 @@ class RenameHandler(ValidatedHandler):
 
     @sessions.requires_edit
     def post(self) -> None:
-        mgr = sessions.get_manager()
+        mgr = get_manager()
         args = parse_raw(self.request.body, Rename)
         filename = args.filename
         LOGGER.debug("Renaming from %s to %s", mgr.filename, filename)
