+++ marimo/_server/api/save_app_config.py
@@ -8,9 +8,10 @@ import tornado.web
 
 from marimo import _loggers
 from marimo._ast import codegen
-from marimo._server import sessions
+from marimo._server import server_utils as sessions
 from marimo._server.api.status import HTTPStatus
 from marimo._server.api.validated_handler import ValidatedHandler
+from marimo._server.sessions import get_manager
 from marimo._utils.parse_dataclass import parse_raw
 
 LOGGER = _loggers.marimo_logger()
@@ -27,7 +28,7 @@ class SaveAppConfigurationHandler(ValidatedHandler):
 
     @sessions.requires_edit
     def post(self) -> None:
-        mgr = sessions.get_manager()
+        mgr = get_manager()
         args = parse_raw(self.request.body, SaveAppConfiguration)
         mgr.update_app_config(args.config)
         # Update the file with the latest app config
