+++ marimo/_server/api/set_cell_config.py
@@ -6,7 +6,7 @@ from typing import Dict
 
 from marimo._ast.cell import CellId_t
 from marimo._runtime import requests
-from marimo._server import sessions
+from marimo._server import server_utils as sessions
 from marimo._server.api.validated_handler import ValidatedHandler
 from marimo._utils.parse_dataclass import parse_raw
