+++ marimo/_server/utils.py
@@ -1,6 +1,8 @@
 # Copyright 2023 Marimo. All rights reserved.
 import os
 import pathlib
+import sys
+from importlib.abc import Traversable
 
 # use spaces instead of a tab to play well with carriage returns;
 # \r\t doesn't appear to overwrite characters at the start of a line,
@@ -16,3 +18,22 @@ def canonicalize_filename(filename: str) -> str:
     if pathlib.Path(filename).suffix != ".py":
         filename += ".py"
     return os.path.expanduser(filename)
+
+
+def import_files(filename: str) -> Traversable:
+    if sys.version_info < (3, 9):
+        from importlib_resources import files as importlib_files
+    else:
+        from importlib.resources import files as importlib_files
+
+    return importlib_files(filename)
+
+
+def initialize_mimetypes() -> None:
+    import mimetypes
+
+    # Fixes an issue with invalid mimetypes on windows:
+    # https://github.com/encode/starlette/issues/829#issuecomment-587163696
+    mimetypes.add_type("application/javascript", ".js")
+    mimetypes.add_type("text/css", ".css")
+    mimetypes.add_type("image/svg+xml", ".svg")
